[TG-TECH][refactor]  refactored devices
