+++ src/app/devices/detail-list-resolve.service.js
@@ -1,0 +1,11 @@
+import { SensorDetailService } from './shared/sensor-detail.service';
+
+export class DeviceListResolveService {
+    constructor(sensorsService: SensorDetailService) {
+        this.sensorsService = sensorsService;
+    }
+
+    resolve() {
+        return this.sensorsService.get();
+    }
+}
