--- src/app/devices/details/index.js
@@ -1,1 +1,0 @@
-export * from './sensor-detail.component.js';
