+++ src/app/devices/device-list/assets/empty-sensors-list.png
