+++ src/app/devices/device-list/device-list.service.js
@@ -1,5 +1,5 @@
-import {Injectable} from '@angular/core';
-import ShHttpService from '../../shared/sh-http/sh-http.service.js';
+import { Injectable } from '@angular/core';
+import { ShHttpService } from '../../shared/sh-http/sh-http.service.js';
 
 @Injectable()
 export class DeviceListService {
