+++ src/app/devices/device-list/device-list.style.scss
