+++ src/app/devices/device-list/device-list.template.html
@@ -1,7 +1,7 @@
 <div class="sub-header row">
   <div id="page-title" class="col s5 sub-header__title"><h2>Devices</h2></div>
   <div class="sub-header__actions col s3 right">
-    <button [routerLink]="['/devices-create']"
+    <button [routerLink]="['/devices/create']"
         class="waves-effect waves-light btn blue-grey darken-1">Add new device</button>
   </div><!-- sub-header__actions -->
 </div><!-- sub-header -->
