--- src/app/devices/index.js
@@ -1,3 +1,0 @@
-import { DeviceList, SensorDetailListResolveService } from './list';
-import { SensorDetail } from './details';
-export { DeviceList, SensorDetail, SensorDetailListResolveService };
