--- src/app/devices/list/index.js
@@ -1,2 +1,0 @@
-export * from './sensor-detail-list-resolve.service.js';
-export * from './device-list.component';
