--- src/app/devices/list/sensor-detail-list-resolve.service.js
@@ -1,12 +1,0 @@
-import SensorDetailService from '../details/sensor-detail.service';
-import { Observable } from 'rxjs/Observable';
-
-export class SensorDetailListResolveService {
-    constructor(sensorsService: SensorDetailService) {
-        this.sensorsService = sensorsService;
-    }
-
-    resolve(): Observable<any> {
-        return this.sensorsService.get();
-    }
-}
