+++ src/app/devices/sensor-details/sensor-detail.style.css
