+++ src/app/devices/sensor-details/sensor-detail.template.html
@@ -5,12 +5,12 @@
     </a>
     <h1>{{ needUpdate ? 'Edit' : 'New' }} device</h1>
 </div>
-<form #sensorForm="ngForm" class="sensor-detail">
+<form class="sensor-detail" #sensorForm="ngForm">
     <div class="sensor-detail__id">
         <label for="sensor-detail__id">Device name <b class="red-text">*</b></label>
         <input [(ngModel)]="sensor.mqttId"
-               ngControl="mqttId"
-               #mqttId="ngForm"
+               name="mqttId"
+               #mqttId="ngModel"
                type="text"
                id="sensor-detail__id"
                placeholder="Enter descriptive name"
@@ -22,8 +22,8 @@
     <div class="sensor-detail__type">
         <label for="sensor-detail__id">Type</label>
         <input [(ngModel)]="sensor.type"
-               ngControl="type"
-               #type="ngForm"
+               name="type"
+               #type="ngModel"
                type="text"
                id="sensor-detail__type"
                placeholder="Enter type"
@@ -32,8 +32,8 @@
     <div class="sensor-detail__description">
         <label for="sensor-detail__id">Description</label>
         <input [(ngModel)]="sensor.description"
-               ngControl="description"
-               #description="ngForm"
+               name="description"
+               #description="ngModel"
                type="text"
                id="sensor-detail__description"
                placeholder="Enter description"
@@ -42,8 +42,8 @@
     <div class="sensor-detail__metrics">
         <label for="sensor-detail__id">Metrics</label>
         <input [(ngModel)]="sensor.metrics"
-               ngControl="metrics"
-               #metrics="ngForm"
+               name="metrics"
+               #metrics="ngModel"
                type="text"
                id="sensor-detail__metrics"
                placeholder="Enter metrics"
@@ -51,7 +51,8 @@
     </div>
     <p class="sensor-detail__status">
         <input [(ngModel)]="sensor.executor"
-               ngControl="executor"
+               name="executor"
+               #executor="ngModel"
                class="filled-in sensor-detail__status-value"
                type="checkbox"
                id="sensor-detail__executor"/>
